requirements.txt
PySide6
SQLAlchemy
PyYAML
docxtpl
python-docx
openapi-core
